feat: allow --inspect-wait when debugging test (#893)
# Motivation
vscode_deno offers a code lens action "Run Test | Debug". However when clicking the "Debug", it triggers a breakpoint which an user did not set.

# Changes made
Unified the logic for selecting the appropriate debugging flag (--inspect-brk or --inspect-wait) based on the behavior of pull request feat: make --inspect-wait default when debugging #775.